JASPAR 2016 Update.
The most current JASPAR DB module. This module was developed to work with the
the JASPAR 2016 release. As of this release the JASPAR DB module will
no longer have a version number attached. The JASPAR DB module without
version number will always be considered the 'current' version. All
JASPAR DB modules with exlicit version numbers will be considered
previous releases tied the the JASPAR DB schema for the related JASPAR
release.

This version of the JASPAR DB module was modified from the JASPAR6.pm
(experimental) module which itself was an incremental update to JASPAR5.pm.
The main change is that for JASPAR 2016, the DB was cleaned up and the
schema changed to be (more) normalized. Specifically, in previous releases
of JASPAR where there were cases of multiple values associated to a
given profile (1-to-many) relationships in tables like MATRIX_ANNOTATION,
MATRIX_SPECIES and MATRIX_PROTEIN the values were (often) stored in a single
comma delimited string. These tables have had their key constraints changed
to allow these 1-to-many relationships to be stored in a  normalized way with
one values per table row. The JASPAR.pm code has been changed to reflect this.
For backward compatibility the code should be able to handle both the
previous comma separated method and the 1-to-n method of storing multiple
values. Note that for cases where there are multiple rows each of which
themselves may contain comma separated string values the code is NOT
robust and should not be relied on.

Please also see the internal code comments for more specifics.